AGENTIC_DQ_VSCODE/requirements.txt
# -------------------------------------------------
# Core Data Processing
# -------------------------------------------------
pandas>=2.0.0
numpy>=1.24.0

# -------------------------------------------------
# LLM & AI
# -------------------------------------------------
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-core>=0.2.0,<0.3.0
langgraph>=0.0.1

# -------------------------------------------------
# Data Quality & Profiling
# -------------------------------------------------
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0,<3.0.0
presidio-analyzer>=0.0.48
presidio-anonymizer>=0.0.48

# -------------------------------------------------
# Web Framework
# -------------------------------------------------
streamlit>=1.28.0

# -------------------------------------------------
# Data Validation & Configuration
# -------------------------------------------------
pydantic>=2.0.0
python-dotenv>=1.0.0

# -------------------------------------------------
# Utilities
# -------------------------------------------------
requests>=2.31.0

# -------------------------------------------------
# Development & Testing
# -------------------------------------------------
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
pylance>=1.1.0

# -------------------------------------------------
# Optional: Monitoring & Logging
# -------------------------------------------------
python-json-logger>=2.0.7

faiss-cpu>=1.7.0
streamlit>=1.28.0,<2.0.0
google-generativeai>=0.3.0,<2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.0
langgraph>=0.0.20,<1.0.0
pydantic==1.10.14  # Critical fix
pydantic-core==2.16.3